Poetry & lots of fixes
