+++ .github/workflows/release.yaml
@@ -1,0 +1,18 @@
+name: Release
+on:
+  release:
+    types: [published]
+jobs:
+  release:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v1
+      with:
+        python-version: 3.7
+        architecture: x64
+    - run: pip install nox==2020.8.22
+    - run: pip install poetry==1.1.4
+    - run: nox
+    - run: poetry build
+    - run: poetry publish --username=__token__ --password=${{ secrets.PYPI_TOKEN }}
