+++ src/lsassy/dumpmethod/comsvcs_stealth.py
