+++ src/lsassy/dumpmethod/dummy.py
