+++ src/lsassy/dumpmethod/dumpertdll.py
