+++ src/lsassy/dumpmethod/idumpmethod.py
