+++ src/lsassy/exec/wmi.py
