+++ src/lsassy/writer.py
