--- tests/__init__.py
@@ -1,5 +1,0 @@
-# Author:
-#  Romain Bentz (pixis - @hackanddo)
-# Website:
-#  https://beta.hackndo.com
-
