+++ tests/test_lsassy.py
@@ -1,0 +1,5 @@
+from lsassy import __version__
+
+
+def test_version():
+    assert __version__ == '0.1.0'
