--- tests/tests_config.py.template
@@ -1,37 +1,0 @@
-# Author:
-#  Romain Bentz (pixis - @hackanddo)
-# Website:
-#  https://beta.hackndo.com
-
-"""
-RENAME THIS FILE TO tests_config.py
-"""
-
-# Include Kerberos authentication tests
-# This test requires to have a valid TGT for a local admin set in KRB5CCNAME
-# See https://github.com/Hackndo/lsassy/wiki/Lsassy-Advanced-Usage#kerberos
-kerberos = False
-
-# Domain controller FQDN or IP, only needed if kerberos set to True
-domain_controller = "192.168.1.101"
-
-# If kerberos is set to True, FQDN of a valid target. IP address otherwise.
-target = "dc01.adsec.local"
-
-# If kerberos is set to True, FQDN of target where LSASS is protected. IP address otherwise. (empty to skip tests)
-protected_target = "protected_server.adsec.local"
-
-# Domain Name
-domain = "adsec.local"
-
-# User with admin rights on target and protected_target
-da_login = "jdoe"
-da_password = "p4ssw0rd"
-
-# User without admin rights on target (empty to skip tests)
-usr_login = "msmith"
-usr_password = "n0rights4U"
-
-# Local tools for dumping methods (empty to skip tests)
-procdump_path = "/opt/procdump.exe"
-dumpert_path = "/opt/dumpert.exe"
